Removed loading screen and extracted piano wrapper from main app
